--- bookstore/src/Components/Categories.js
@@ -1,10 +1,0 @@
-const Category = () => {
-    
-    return(
-        <div>
-        <button type="button">Check status</button>
-    </div>
-    )
-};
-
-export default Category;
