--- bookstore/src/index.js
@@ -1,13 +1,0 @@
-import React from "react";
-import { BrowserRouter } from 'react-router-dom';
-import ReactDOM  from "react-dom";
-import App from './App';
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-    <React.StrictMode>
-        <BrowserRouter>
-           <App />
-        </BrowserRouter>
-    </React.StrictMode>,
-);
