+++ src/App.css
@@ -1,0 +1,3 @@
+body {
+  width: 100%;
+}
